Add css.TailwindCSS
Closes #12618
Closes #12620
